move network related files from `pkg/util/cloudproviders` to own package
