+++ pkg/util/cloudproviders/network/network.go
@@ -3,7 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package cloudproviders
+// Package network provides utilities around cloud provider networking.
+package network
 
 import (
 	"context"
