+++ pkg/util/cloudproviders/network/network_mock.go
@@ -5,7 +5,7 @@
 
 //go:build test
 
-package cloudproviders
+package network
 
 import (
 	"testing"
