+++ pkg/util/cloudproviders/network/network_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2025-present Datadog, Inc.
 
-package cloudproviders
+package network
 
 import (
 	"context"
